State machine correction for sawtooth-raft
A leader node in sawtooth-raft asks validator to commit a block
when majority of the follower nodes replicated the log entry.
However follower nodes may lose the log commit confirmation
from the leader node. Majority of the follower nodes form a
group and elect a new leader.

A log entry is safe committed as per raft if majority of the
nodes have the log stored and at least one entry from the
leaders term is also majority stored.

If majority of the nodes continue adding blocks, the old
leader has to replace the log entry. This is taken care by raft.
However block already added to the blockchain shall also get
replaced with the new one.

Order of the nodes added to blockchain is controlled by deque.

Signed-off-by: S m, Aruna <aruna.s.m@intel.com>